make BaseModal component
